refactor: MemberRecommendationDto로 수정 및 위치 변경
