+++ src/main/java/com/bookbla/americano/domain/member/repository/custom/MemberRepositoryCustom.java
@@ -2,9 +2,8 @@ package com.bookbla.americano.domain.member.repository.custom;
 
 import com.bookbla.americano.domain.member.controller.dto.request.MemberBookProfileRequestDto;
 import com.bookbla.americano.domain.member.controller.dto.response.BookProfileResponse;
-import com.bookbla.americano.domain.member.controller.dto.response.MemberRecommendationResponse;
+import com.bookbla.americano.domain.member.service.dto.MemberRecommendationDto;
 import com.bookbla.americano.domain.member.repository.entity.Member;
-import com.bookbla.americano.domain.member.repository.entity.MemberBook;
 import com.bookbla.americano.domain.postcard.repository.entity.Postcard;
 
 import java.util.List;
@@ -13,5 +12,5 @@ public interface MemberRepositoryCustom {
 
     List<BookProfileResponse> getAllMembers(Long memberId, MemberBookProfileRequestDto requestDto);
 
-    List<Member> getRecommendationMembers(MemberRecommendationResponse memberRecommendationResponse, List<Postcard> postcards);
+    List<Member> getRecommendationMembers(MemberRecommendationDto memberRecommendationResponse, List<Postcard> postcards);
 }
